\name{choose_input_licor_files}

\alias{choose_input_licor_files}

\title{Choosing Licor Excel files}

\description{
  Opens up a window to choose Licor Excel files. Only available in interactive
  R sessions running in Microsoft Windows. The output from this function is
  typically passed to \code{\link{read_licor_file}}.
}

\usage{
  choose_input_licor_files()
}

\value{
  A character vector of full file paths.
}
