\name{PhotoGEA}

\alias{PhotoGEA}
\alias{photogea}

\title{The PhotoGEA R package}

\description{
  \strong{PhotoGEA} (short for \strong{photo}synthetic \strong{g}as
  \strong{e}xchange \strong{a}nalysis) is an R package that provides a suite of
  tools for loading, processing, and analyzing photosynthetic gas exchange data.
  See Lochocki, Salesse-Smith, & McGrath (2025) [\doi{10.1111/pce.15501}] for
  more information.

  The best way to learn about using PhotoGEA is to visit the
  \href{https://eloch216.github.io/PhotoGEA/index.html}{PhotoGEA website} and
  click the
  \href{https://eloch216.github.io/PhotoGEA/articles/PhotoGEA.html}{Get Started}
  link in the top menu bar. The website includes documentation for all the
  functions and data sets included in the package, as well as articles that
  describe its general features and several important use cases.
}
