\name{deprecated}

\alias{deprecated}
\alias{read_tdl_file}

\title{Reading a CR3000 data file}

\description{
  Deprecated functions that will be fully removed in future releases. Each of
  these functions will produce an error when called that will redirect the user
  to a suitable replacement.
}

\usage{
  read_tdl_file(\dots)
}

\arguments{
  \item{\dots}{
    Additional arguments; currently unused.
  }
}
